+++ docs/source/changelog.rst
@@ -10,6 +10,7 @@ v26.1.10-dev
 
 - Receive all integration-related gateway events.
 - :attr:`~discord.objects.Channel.application_id` is now also :attr:`~discord.utils.Nullable`.
+- Added :class:`~discord.events.IntegrationDeleteEvent` event.
 
 
 **Documentation Changes**